+++ src/main/java/com/example/spot/service/memberstudy/MemberStudyCommandService.java
@@ -4,13 +4,13 @@ import com.example.spot.web.dto.memberstudy.request.StudyQuizRequestDTO;
 import com.example.spot.web.dto.memberstudy.response.StudyQuizResponseDTO;
 import com.example.spot.web.dto.memberstudy.response.StudyTerminationResponseDTO;
 import com.example.spot.web.dto.memberstudy.response.StudyWithdrawalResponseDTO;
-import com.example.spot.web.dto.study.request.ScheduleRequestDTO;
-import com.example.spot.web.dto.study.request.StudyPostCommentRequestDTO;
-import com.example.spot.web.dto.study.request.StudyPostRequestDTO;
-import com.example.spot.web.dto.study.response.ScheduleResponseDTO;
+import com.example.spot.web.dto.memberstudy.request.ScheduleRequestDTO;
+import com.example.spot.web.dto.memberstudy.request.StudyPostCommentRequestDTO;
+import com.example.spot.web.dto.memberstudy.request.StudyPostRequestDTO;
+import com.example.spot.web.dto.memberstudy.response.ScheduleResponseDTO;
 import com.example.spot.web.dto.study.response.StudyApplyResponseDTO;
-import com.example.spot.web.dto.study.response.StudyPostCommentResponseDTO;
-import com.example.spot.web.dto.study.response.StudyPostResDTO;
+import com.example.spot.web.dto.memberstudy.response.StudyPostCommentResponseDTO;
+import com.example.spot.web.dto.memberstudy.response.StudyPostResDTO;
 
 public interface MemberStudyCommandService {
