+++ src/main/java/com/example/spot/service/memberstudy/MemberStudyCommandServiceImpl.java
@@ -19,13 +19,13 @@ import com.example.spot.web.dto.memberstudy.request.StudyQuizRequestDTO;
 import com.example.spot.web.dto.memberstudy.response.StudyQuizResponseDTO;
 import com.example.spot.web.dto.memberstudy.response.StudyTerminationResponseDTO;
 import com.example.spot.web.dto.memberstudy.response.StudyWithdrawalResponseDTO;
-import com.example.spot.web.dto.study.request.StudyPostCommentRequestDTO;
-import com.example.spot.web.dto.study.request.StudyPostRequestDTO;
+import com.example.spot.web.dto.memberstudy.request.StudyPostCommentRequestDTO;
+import com.example.spot.web.dto.memberstudy.request.StudyPostRequestDTO;
 import com.example.spot.web.dto.study.response.StudyApplyResponseDTO;
-import com.example.spot.web.dto.study.request.ScheduleRequestDTO;
-import com.example.spot.web.dto.study.response.ScheduleResponseDTO;
-import com.example.spot.web.dto.study.response.StudyPostCommentResponseDTO;
-import com.example.spot.web.dto.study.response.StudyPostResDTO;
+import com.example.spot.web.dto.memberstudy.request.ScheduleRequestDTO;
+import com.example.spot.web.dto.memberstudy.response.ScheduleResponseDTO;
+import com.example.spot.web.dto.memberstudy.response.StudyPostCommentResponseDTO;
+import com.example.spot.web.dto.memberstudy.response.StudyPostResDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
